fix(kafka-connect-jms): Ensure JMSSessionProvider only creates JMS Consumers and JMS Producers when appropriate.
JMS Consumers are only created when in 'source' mode and JMS Producers are only created when in 'sink' mode.

Refactored some of the supporting test code so that the ActiveMQ brokers are started on a free port if none is specified at construction time. This prevents port collisions between tests and allows tests to potentially be run in paralle.

Fixes #330