feat(MQTT): replicate shared subscription to all tasks if enabled
With this new option, the messages published to shared subscriptions are
distributed over all tasks. Therefore, the work generated by a KCQL
instruction with a shared subscription is now divided over all tasks. It
is the job of the MQTT broker to load balance the messages to the
connected tasks.