MongoDB writer
Existing code was incorrectly mapping the keys and values. From KCQL it was not returning the full path for a projection and that caused issue with the projection.

Also the key extractions was wrongfully done after the projection.

The change allows for the full field projection list to be returned and the key projection now has the original source