Fix division by zero bug (#944)
* Fix the bug where the ConnectorTaskId had the task number and the max tasks values reversed. This created a runtime exception on PartitionHasher (division by zero).

Removes the ConnectorTaskId trait and creates a case class. There was no reason to default when the name  or TASK_INDEX properties were missing.

Adds the unit test around the ConnectorTaskId to check the values for max tasks and task number.

* Bring back the .show to use the name for ConnectorTaskId

Removes invalid unit test in LocalStagingAreaTest

* Fix the compilation errors on the tests

* Fix the IT tests

Fix the ConnectorTaskId.ownsDir to avoid adding + 1 on the hasher (not its responsibility). Adds tests for the ownsDir

Adds unit tests for TaskDistributor

---------

Co-authored-by: stheppi <jenkins@lenses.io>