ReaderCreator.scala was constructed to call create. The code was moved to ResultReader as a companion object.
Avoid a copy of excluded partitions in PartitionSearcher

ReaderManager gets the connectorTaskId as ctor param.

ReaderManager initialisation gets the Ref for ReaderManagerState and avoids the unsafeRunSync in its ctor. (More needs to be done to avoid the occasional unsafeRunSync)

S3Location does not have pathOrError. It reduces the complexity of navigating a few functions to get the result versus having a leftMap where pathOrError was called.