ReaderManager has been rewritten to avoid so many states. A file is read and once it completes it moves to the next file, if there is one. It reduces the code complexity of having a few state types. Furthermore, the code was changed to use the effect type IO.
S3SourceFileQueue has dropped the markComplete and init. MarkComplete is redundant if the class was to behave as a queue. Once a file is picked is removed from the internals and it does not require an external call to read. The init function is not ideal to allow it to be called at any point. It has been replaced with a specific builder which sets the initial state to the one initial file.

S3SourceTaskState has been updated to initialise the Ref[IO, T] and avoid the unsafe run sync call.