Feat/improvements rebase (#947)
* S3 Source - Support for ordering by date
Unit testing

* Removes the function to return the InputStream / Outputstream. There is no reason to provide the function since the instances are being used immediately. This allows to have the for comprehension containing the instruction to create the input/output.

Removes the Open/Close state. The implementation requires specific methods to throw an exception. The new code relies on Option[ofstate]

Removes AuthResource.scala since it has 2 indirections to call new AwsS3ClientCreator.createS3Client. The trait is not used, and it creates an AuthResources to call aws.

Removes ThrowableEither.scala and moves the left side of either to throwable to avoid having the class.

Small improvements here and there avoiding redundant calls in for-comprehension

* ConnectorTaskId.fromProps is returning a Throwable, which avoids a leftMap on S3SourceTaskState

* Fix the compilation errors

* ReaderCreator.scala was constructed to call create. The code was moved to ResultReader as a companion object.

Avoid a copy of excluded partitions in PartitionSearcher

ReaderManager gets the connectorTaskId as ctor param.

ReaderManager initialisation gets the Ref for ReaderManagerState and avoids the unsafeRunSync in its ctor. (More needs to be done to avoid the occasional unsafeRunSync)

S3Location does not have pathOrError. It reduces the complexity of navigating a few functions to get the result versus having a leftMap where pathOrError was called.

* ReaderManager has been rewritten to avoid so many states. A file is read and once it completes it moves to the next file, if there is one. It reduces the code complexity of having a few state types. Furthermore, the code was changed to use the effect type IO.

S3SourceFileQueue has dropped the markComplete and init. MarkComplete is redundant if the class was to behave as a queue. Once a file is picked is removed from the internals and it does not require an external call to read. The init function is not ideal to allow it to be called at any point. It has been replaced with a specific builder which sets the initial state to the one initial file.

S3SourceTaskState has been updated to initialise the Ref[IO, T] and avoid the unsafe run sync call.

* Fix integration tests compilation errors

Fix ConnectorTaskIdTest

* Fix the parquet reader. The code change omitted the skip bytes on resync.

* Remove unused import

* Avoid NPE when the sink/source task context is null, which happens a lot in the integration tests

---------

Co-authored-by: David Sloan <david.sloan@lenses.io>
Co-authored-by: stheppi <jenkins@lenses.io>