Description
AwsS3DirectoryLister removes the trailing /. If not, it won't identify the bucket:prefix as a partition if that's the case. There was also a bug decreasing the levels

Introduces LineStartLineEndReader to cater for the scenario where a line is a merge of multiple lines where the first and last match a setting.

Removes the skipLines in the reader. Single Responsibility Principles dictates that. Also, it was wrong, because one line at source is not the same as one line in the reader.