Fix/s3 source (#954)
* Description

AwsS3DirectoryLister removes the trailing /. If not, it won't identify the bucket:prefix as a partition if that's the case. There was also a bug decreasing the levels

Introduces LineStartLineEndReader to cater for the scenario where a line is a merge of multiple lines where the first and last match a setting.

Removes the skipLines in the reader. Single Responsibility Principles dictates that. Also, it was wrong, because one line at source is not the same as one line in the reader.

* Introduces an option for controlling the buffer size for reading with prefix-suffix

Rewrites the PrefixSuffixReader to use a buffer and not rely on reading a line. There can be large xml docs on one line.

* Improve the unit tests names for TextUtils

* Extending the log

* Fixing the code format

* Fixing the errors on the unit tests.

* Fix the error on TextUtilsTest

* Addressing PR improvement suggestions.

* Fix the integration tests

---------

Co-authored-by: stheppi <jenkins@lenses.io>