Control partitions discovery
Not all connectors need to have dynamic partition discovery. Some of the partitions are static. Running the partitions discovery once reduces the CPU and traffic to S3.

The code introduces a new property, of boolean type, to control the search. If set to true, the search is continuous, otherwise it is a one of.

PollLoop is extended with a new method which runs an IO until it is not failing. This is done to handle scenarios where there are intermittent failures (i.e. connectivity issues) and therefore the connector partition discovery can happen.