Control partitions discovery (#959)
* Control partitions discovery

Not all connectors need to have dynamic partition discovery. Some of the partitions are static. Running the partitions discovery once reduces the CPU and traffic to S3.

The code introduces a new property, of boolean type, to control the search. If set to true, the search is continuous, otherwise it is a one of.

PollLoop is extended with a new method which runs an IO until it is not failing. This is done to handle scenarios where there are intermittent failures (i.e. connectivity issues) and therefore the connector partition discovery can happen.

* fix the unit test

* Move away from Thread.sleep and leverage eventually

* Fix the unit tests, making sure the background task for partition discovery is closed

* Avoid background task hanging around and stopping the integration tests

---------

Co-authored-by: stheppi <jenkins@lenses.io>