Refactors the readers to avoid the conversion to SourceRecord in the reader - keeps their single purpose.
Introduces the Converter to take care of the conversion to SourceRecord.

Introduces DelegateIteratorS3StreamReader to be responsible for calculating the record index such that the readers do not repeat the functionality.

It was identified two serious bugs on the source because of the -1 based index in the readers current record index. This means index 0 is 1 record. When the connector is restarted the result is the last record is processed again.

Another issue was the offset where the Instant.now was uses as the file last modified as opposed of the actual file/object lastModified