Support full backup on S3 sink (#975)
* First commit

* Extends the MessageDetail to contain the offset.

Renames the MessageDetail fields to key/value

Introduces configuration options to store key/headers/metadata as well. The default to true

S3SinkTask:put avoids computing the debug log values when it is not required

Topic class is extending AnyVal to avoid extra runtime allocations.

* Adds the logic for enabling the envelope.

For this version, to keep backwards compatibility the rules are as follow:

No properties set, it defaults to no-envelope, and all fields set to true.

store.envelope=true if present it enables envelope and all fields

to discard any of the envelope fields, all fields need to be specified.

* Add unit tests for MessageTransformer. Build the envelope without assuming Value is always present.

Fix some compilation errors in integration test

Rename TestSampleSchemaAndData.scala to SampleData

* Handling Array and Map for the message transformer

* Avoids wrapping Array and Maps with SinkData. Avoids the object allocation and CPU cycles.

It also reduces the need for convertes when using JsonConverter

Makes SinkData value an Any to avoid boxing java primitives

Fixes the compilation errors in the integration tests

* Makes the MessageDetail Key a SinkData not an Option[SinkData]. Therefore the MessageDetail represents the incoming Kafka message

* Adds validation for storage format when envelope is used. Text, CSV and Bytes is not allowed when envelope is used

* Introduces a Transformer interface and converts the MessageTransformer.scala to EnvelopeWithSchemaTransformer.

Introduces a sequence transformer as well as the TopicsTransformers.

Enables the message transformation in the S3WriterManager

* Fix compilation errors

* Fixing unit tests failures.

* Fix the unit test on ValueToSinkDataConverterTest

* Fixing compilation error

* Remove the trace log from ResultReader. That could lead to huge log entries when set to trace, one for each record read.

Changed the IT test log level to DEBUG.

* Fix the integration tests

* Adding the Schema less transformer used for JSON storage

* Revert ToAvroDataConverter to use  ByteBuffer.wrap for Array[Byte]

* Add the integration tests for AVRO and PARQUET envelope

* Introduces the escape new line for JSON,CSV and text.

* Revert back the escape transformer. It was not needed

* Refactor the code to remove SourceData.scala.

S3FormatStreamReader.scala was renamed to S3StreamReader and made an Iterator[SourceRecord].

* Refactors the readers to avoid the conversion to SourceRecord in the reader - keeps their single purpose.

Introduces the Converter to take care of the conversion to SourceRecord.

Introduces DelegateIteratorS3StreamReader to be responsible for calculating the record index such that the readers do not repeat the functionality.

It was identified two serious bugs on the source because of the -1 based index in the readers current record index. This means index 0 is 1 record. When the connector is restarted the result is the last record is processed again.

Another issue was the offset where the Instant.now was uses as the file last modified as opposed of the actual file/object lastModified

* Fix compilation errors.

Json envelope adds two fields for key and value if they are binary, since the Json serializer will base64 encode them

* Add the source envelope converters for Avro/Parquet and Json.

For JSON to convert back to arrays it uses these two fields:keyIsArray/valueIsArray

* Add the integration tests for the S3 source for Avro/Parquet/Json when envelope is involved

* Remove unused import

* Fixing broken tests

* Fix the offset test. The code starts from 0 hence the test is counting more than it should.

* Moving to 5.0 version

* Remove redundant code MapToStructConverter.scala.

WrappedComplexTypeExtractor handles java.util.List

---------

Co-authored-by: stheppi <jenkins@lenses.io>