introduces the S3 sink for consumer groups offsets.
The sink stores the latest offsets for a group-topic-partition under a key: $bucket/[$prefix]/$group/$topic/$partition

Since eventually the latest offset is written, the connector does not need to track the offsets for a given key, thus reducing the complexity.

Consumer offsets topic contains OffsetsKey messages and GroupMetadata. The latter is ignored since it deals with the group information.