[OPS-2283] CICD: Add AV checks before release
This commit adds antivirus checks via ClamAV on release builds.

The scan runs in two jobs, the first one refreshes the AV database and caches
it, then checks the repo itself for malicious files. Then the second job, runs
in a matrix and scans the build files including their dependencies.
We are being cautious and enable most of ClamAV's features.

Signed-off-by: Stefan Bocutiu <stefan.bocutiu@gmail.com>
Signed-off-by: Marios Andreopoulos <opensource@andmarios.com>