LC-195: Introduce an 8.0 Compatibility Setting to Streamline Future Migrations
ome customers are onboarding on 7.0 which will bring changes to the KCQL statement. For 8.0 there will be a new property to address faster seeks on the source.

By incorporating this property now, users can seamlessly integrate it into their workflows, preempting the need for a subsequent migration to the upcoming version 8.0.