Enhancements for Source Line-Start-End Functionality (LC-196) (#1184)
* Enhancements for Source Line-Start-End Functionality (LC-196)

This PR addresses a specific challenge encountered by users of the S3 source functionality. In cases where an external producer terminates a file without the end message mark. This ends up losing data.

To mitigate this issue, the PR introduces a new property entry for KCQL to signal the unterminated message to be considered.

* Fixes the compilation error and adds a test for initialising the StartEndLineReadTextMode

---------

Co-authored-by: stheppi <jenkins@lenses.io>