LC-137 Key names
This is a rollback of some of the changes around key names in order to support data reloads from specific timestamp.

The change reverts back the TopicPartitionOffsetFileNamer since it is not impacted by the data reload.

For OffsetFileNamer the two versions (v0 and v1) were removed since the change is backwards compatible.