Staggered commit (#150)
In the original implementation of the Cloud Data Lake Sinks (AWS S3, Azure Data Lake and GCP Storage), when one writer commits, then all writers are committed in sync.  Whilst this maintains predictability in terms of logging and commit logic, unfortunately this is sub-optimal for multiple reasons:
* This leads to many small files building up in the data lake, meaning it can be inefficient.
* More upload operations are triggered, leading to chattier networks.
* Often the limits specified in the flush configuration are not reached for most files, leading to users questioning the behaviour.

This PR changes this behaviour so that when a writer is ready to flush, only writers related to that topic/partition will be committed (instead of all).

If custom partitioning is used, it could still lead to some of the previous behaviour of small files as Topic/Partition writers need to be committed in tandem.  However, for most cases this will ensure that Topic/Partitions reach their trigger point instead of being committed in cascade by other writers.

In terms of the code, there have been some small changes:

* The logic for deciding which writers to commit has been moved out of `WriterManager` into a new unit called `WriterCommitManager`.  This is to enable separation of concerns and simpler testability.  Unit tests have been added for this logic.
* Some simplification of the logging, logic and removal of some of the previous methods.