Avoid filtering by lastSeenFile where a post process action is configured (#192)
Avoid using lastModified time for filtering in cases where a post process action is configured.

In the CloudSourceFileQueue in cases where a post process action is configured, we no configure the lastSeenFile when retrieving the new list of files. This means the files in the BatchLister will no longer be filtered, and consequently no files will be skipped by the source.