HTTP JMX Fixes (#199)
* HTTP JMX Fixes

The code makes it clear there's one metrics across a given task.

The previous histogram implementation was not handling the non-thread safe Histogram implementation. With this change the code leverages the Recorder, which is the recommended option for time window histograms. The metrics are also moved to be computed to a 5 minutes interval.

A fixed scheduler sets the reset of the histogram on the 5 minutes mark, and to avoid passing the time window before the values are updated, the scheduler also updates the percentiles every 30 seconds

* Removes the double registering of the mbean

---------

Co-authored-by: stheppi <jenkins@lenses.io>