HTTP extractor (#198)
* HTTP extractor

A bug was reported that referencing the entire record value yields an exception when the payload is a complex type. In this case a HashMap.

The changes allow the return of the value when the extractor has no field to extract (ie. extract the full value)

* Remove the comment

* Improve code readability

Use the Scala Map to Java map for code readability

Co-authored-by: David Sloan <33483659+davidsloan@users.noreply.github.com>

* Fix the unit test after committing the code change suggestion

* Fix the unit test expecting error on unknown type and no path

---------

Co-authored-by: stheppi <jenkins@lenses.io>
Co-authored-by: David Sloan <33483659+davidsloan@users.noreply.github.com>