LC-292 Duplicate connectors (#218)
* LC-292 Duplicate connectors

Users might want to run two different connectors, on two different Kafka clusters, reading from the same topic and storing in the same bucket. When this happens, it can lead to data loss because the two connectors can generate the same object key.

With this change, a new properties setting is added: key.suffix to allow the user to add a unique value to each connector generated keys to avoid the collision.

* Fix the integration tests

---------

Co-authored-by: stheppi <jenkins@lenses.io>