LC-303 Lock file location (#278)
* LC-303 Lock file location

The lock file location has to use the connector name to avoid overlapping connectors. If that happens it leads to corrupted state and corrupted data. Of course, if there are more than 1 Connect cluster used targeting the same datalake bucket and using the same topic name and connector name, the connector cannot help. It is expected the user to be responsible for the setup (docs highlight the scenario).

The change prefixes with the connector name. A migration is required once the change is applied to move the .locks to a parent "folder" containing the connector name.

* Move the tests logs to INFO

* fix the unit test

* fix the unit test

* fix the unit test

---------

Co-authored-by: stheppi <jenkins@lenses.io>