Fix/lc 303 lock file path incorrect (#280)
* LC-303 Lock file location

The lock file location has to use the connector name to avoid overlapping connectors. If that happens it leads to corrupted state and corrupted data. Of course, if there are more than 1 Connect cluster used targeting the same datalake bucket and using the same topic name and connector name, the connector cannot help. It is expected the user to be responsible for the setup (docs highlight the scenario).

The change prefixes with the connector name. A migration is required once the change is applied to move the .locks to a parent "folder" containing the connector name.

* fix the unit test

* The lock file migration should contain the connector name to avoid scenarios where two connectors deal with the same topic and thus corrupts their state leading to data loss.

But the change made leads to a migration from the old lock path to the new one. To avoid manual migration step, the code looks if the previous lock location exists, and if it does, it moves it to the new location.

* Apply the linter

* Fix the unit test

---------

Co-authored-by: stheppi <jenkins@lenses.io>