LC-304 S3 sink broken on non-directory buckets (#279)
When the temp object is deleted, it uses the ifMatch condition on the etag to ensure it has not been changed. But the temp object key already uses a UUID, so different tasks would not overlap in the temp object name. Thus removing the ifMatch condition not only improves performance, but also removes the blocker for non-directory buckets

Co-authored-by: stheppi <jenkins@lenses.io>